+++ src/delfino_docker/__init__.py
@@ -10,7 +10,8 @@ from delfino.models import AppContext
 from delfino.terminal_output import print_header
 from delfino.utils import ArgsList
 from delfino.validation import assert_package_manager_is_known, assert_pip_package_installed, pyproject_toml_key_missing
-from delfino_core.config import CorePluginConfig, Dockerhub, pass_plugin_app_context
+
+from delfino_docker.config import Dockerhub, DockerPluginConfig, pass_plugin_app_context
 
 try:
     from packaging.version import Version
@@ -45,7 +46,7 @@ def _docker_build(
     _flags = list(flags)
     _flags.extend(["--platform", platform])
     if tag != "latest" and push:
-        _flags.extend(["--cache-to", f"type=registry,ref={dockerhub.username}/{project_name}"])
+        _flags.extend(["--cache-to", f"type=registry,ref={dockerhub.dockerhub_username}/{project_name}"])
 
     run(
         [
@@ -53,9 +54,9 @@ def _docker_build(
             "buildx",
             "build",
             "--cache-from",
-            f"type=registry,ref={dockerhub.username}/{project_name}",
+            f"type=registry,ref={dockerhub.dockerhub_username}/{project_name}",
             "--tag",
-            f"{dockerhub.username}/{project_name}:{tag}",
+            f"{dockerhub.dockerhub_username}/{project_name}:{tag}",
             *_flags,
             ".",
         ],
@@ -73,7 +74,7 @@ def _docker_build(
 )
 @click.option("--serialized", is_flag=True, help="Do not build multiple platforms concurrently.")
 @pass_plugin_app_context
-def docker_build(app_context: AppContext[CorePluginConfig], push: bool, serialized: bool):
+def docker_build(app_context: AppContext[DockerPluginConfig], push: bool, serialized: bool):
     """Build and push a docker image."""
     plugin_config = app_context.plugin_config
     dockerfile = Path("Dockerfile")
@@ -85,7 +86,7 @@ def docker_build(app_context: AppContext[CorePluginConfig], push: bool, serializ
     assert_pip_package_installed("packaging")
     assert dockerfile.exists() and dockerfile.is_file(), "File 'Dockerfile' not found but required by this command."
     assert app_context.pyproject_toml.tool.poetry, pyproject_toml_key_missing("tool.poetry")
-    assert plugin_config.dockerhub, pyproject_toml_key_missing("tool.delfino.dockerhub")
+    assert plugin_config.docker_build, pyproject_toml_key_missing("tool.delfino-docker.docker_build")
 
     print_header("Running Docker build")
 
@@ -98,32 +99,37 @@ def docker_build(app_context: AppContext[CorePluginConfig], push: bool, serializ
     python_version = Version(poetry.dependencies["python"].strip("<>=~^"))
     flags.extend(["--build-arg", f"PYTHON_VERSION={python_version.major}.{python_version.minor}"])
 
-    dockerhub = plugin_config.dockerhub
+    command_config = plugin_config.docker_build
 
     if push:
         dockerhub_password = getenv("DOCKERHUB_PERSONAL_ACCESS_TOKEN")
         if not dockerhub_password:
             click.secho("⚠ The 'DOCKERHUB_PERSONAL_ACCESS_TOKEN' environment variable is not set.", fg="yellow")
             dockerhub_password = getpass("Dockerhub Personal Access Token: ")
-        run(f"docker login --username {dockerhub.username} --password {dockerhub_password}", on_error=OnError.EXIT)
+        run(
+            f"docker login --username {command_config.dockerhub_username} --password {dockerhub_password}",
+            on_error=OnError.EXIT,
+        )
 
     if getenv("CI"):  # https://circleci.com/docs/2.0/env-vars/#built-in-environment-variables
         flags.extend(["--progress", "plain"])
 
-    joined_build_platforms = ",".join(dockerhub.build_for_platforms)
+    joined_build_platforms = ",".join(command_config.build_for_platforms)
     if serialized:
-        build_platforms: List[str] = dockerhub.build_for_platforms
+        build_platforms: List[str] = command_config.build_for_platforms
     else:
         build_platforms = [joined_build_platforms]
-        _install_emulators(dockerhub.build_for_platforms)
+        _install_emulators(command_config.build_for_platforms)
 
     # If serialized build is selected, run build individually. Results are cached so the second
     # build below will only push and not build again.
     for platform in build_platforms:
         if serialized:
             _install_emulators([platform])
-        print_header(f"Build {dockerhub.username}/{project_name}:latest for {platform}", level=2, icon="🔨")
-        _docker_build(project_name, dockerhub, flags, platform)
+        print_header(
+            f"Build {command_config.dockerhub_username}/{project_name}:latest for {platform}", level=2, icon="🔨"
+        )
+        _docker_build(project_name, command_config, flags, platform)
 
     if not push:
         return
@@ -133,7 +139,7 @@ def docker_build(app_context: AppContext[CorePluginConfig], push: bool, serializ
     # While `docker buildx build` supports multiple `--tag` flags, push of them fails to expose
     # all architectures in `latest`. Multiple pushes fix this.
     for tag in [project_version, "latest"]:
-        print_header(f"Push {dockerhub.username}/{project_name}:{tag}", level=2, icon="🔨")
-        _docker_build(project_name, dockerhub, flags, joined_build_platforms, tag, push)
+        print_header(f"Push {command_config.dockerhub_username}/{project_name}:{tag}", level=2, icon="🔨")
+        _docker_build(project_name, command_config, flags, joined_build_platforms, tag, push)
 
     run("docker logout", on_error=OnError.EXIT)